+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/channel/ControllerEvent.java
@@ -29,7 +29,7 @@ import com.google.auto.value.AutoValue;
 @AutoValue
 public abstract class ControllerEvent implements ChannelEvent {
 
-    public static ControllerEvent create(int index, int tick, int channel, int controller, int value) {
+    public static ControllerEvent create(long index, int tick, int channel, int controller, int value) {
         return new AutoValue_ControllerEvent(index, tick, channel, controller, value);
     }
