+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/channel/PitchBendEvent.java
@@ -29,7 +29,7 @@ import com.google.auto.value.AutoValue;
 @AutoValue
 public abstract class PitchBendEvent implements ChannelEvent {
 
-    public static PitchBendEvent create(int index, int tick, int channel, int pitch) {
+    public static PitchBendEvent create(long index, int tick, int channel, int pitch) {
         return new AutoValue_PitchBendEvent(index, tick, channel, pitch);
     }
