+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/channel/ProgramChangeEvent.java
@@ -29,7 +29,7 @@ import com.google.auto.value.AutoValue;
 @AutoValue
 public abstract class ProgramChangeEvent implements ChannelEvent {
 
-    public static ProgramChangeEvent create(int index, int tick, int channel, int program) {
+    public static ProgramChangeEvent create(long index, int tick, int channel, int program) {
         return new AutoValue_ProgramChangeEvent(index, tick, channel, program);
     }
