+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/meta/CuePointEvent.java
@@ -30,7 +30,7 @@ import com.techshroom.midishapes.midi.event.MidiEvent;
 @AutoValue
 public abstract class CuePointEvent implements MidiEvent {
 
-    public static CuePointEvent create(int index, int tick, int channel, String cue) {
+    public static CuePointEvent create(long index, int tick, int channel, String cue) {
         return new AutoValue_CuePointEvent(index, tick, channel, cue);
     }
