+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/meta/LyricEvent.java
@@ -30,7 +30,7 @@ import com.techshroom.midishapes.midi.event.MidiEvent;
 @AutoValue
 public abstract class LyricEvent implements MidiEvent {
 
-    public static LyricEvent create(int index, int tick, int channel, String lyric) {
+    public static LyricEvent create(long index, int tick, int channel, String lyric) {
         return new AutoValue_LyricEvent(index, tick, channel, lyric);
     }
