+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/meta/MarkerEvent.java
@@ -30,7 +30,7 @@ import com.techshroom.midishapes.midi.event.MidiEvent;
 @AutoValue
 public abstract class MarkerEvent implements MidiEvent {
 
-    public static MarkerEvent create(int index, int tick, int channel, String marker) {
+    public static MarkerEvent create(long index, int tick, int channel, String marker) {
         return new AutoValue_MarkerEvent(index, tick, channel, marker);
     }
