+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/meta/SetTempoEvent.java
@@ -35,7 +35,7 @@ public abstract class SetTempoEvent implements MidiEvent {
 
     public static final long MICROSECONDS_PER_MINUTE = TimeUnit.MINUTES.toMicros(1);
 
-    public static SetTempoEvent createBpm(int index, int tick, int channel, double bpm) {
+    public static SetTempoEvent createBpm(long index, int tick, int channel, double bpm) {
         // have bpm == beats per minute
         // have MPM == micros per minute
         // want: micros per beat
@@ -43,7 +43,7 @@ public abstract class SetTempoEvent implements MidiEvent {
         return create(index, tick, channel, (int) (MICROSECONDS_PER_MINUTE / bpm));
     }
 
-    public static SetTempoEvent create(int index, int tick, int channel, int tempo) {
+    public static SetTempoEvent create(long index, int tick, int channel, int tempo) {
         return new AutoValue_SetTempoEvent(index, tick, channel, tempo);
     }
