+++ MidiShapes2/src/main/java/com/techshroom/midishapes/midi/event/meta/TextEvent.java
@@ -30,7 +30,7 @@ import com.techshroom.midishapes.midi.event.MidiEvent;
 @AutoValue
 public abstract class TextEvent implements MidiEvent {
 
-    public static TextEvent create(int index, int tick, int channel, String text) {
+    public static TextEvent create(long index, int tick, int channel, String text) {
         return new AutoValue_TextEvent(index, tick, channel, text);
     }
